requirements.txt
Django==1.11.5
djangorestframework==3.6.4
django-fontawesome
